+++ Common/src/main/java/fuzs/puzzleslib/api/biome/v1/GenerationSettingsContext.java
@@ -7,9 +7,16 @@ import net.minecraft.world.level.levelgen.GenerationStep;
 import net.minecraft.world.level.levelgen.carver.ConfiguredWorldCarver;
 import net.minecraft.world.level.levelgen.placement.PlacedFeature;
 
-public
-
-interface GenerationSettingsContext {
+/**
+ * The modification context for the biomes generation settings.
+ *
+ * <p>Mostly copied from Fabric API's Biome API, specifically <code>net.fabricmc.fabric.api.biome.v1.BiomeModificationContext$WeatherContext</code>
+ * to allow for use in common project and to allow reimplementation on Forge using Forge's native biome modification system.
+ *
+ * <p>Copyright (c) FabricMC
+ * <p>SPDX-License-Identifier: Apache-2.0
+ */
+public interface GenerationSettingsContext {
     /**
      * Removes a feature from one of this biomes generation steps, and returns if any features were removed.
      */
@@ -119,7 +126,15 @@ interface GenerationSettingsContext {
         return this.removeCarver(step, BuiltinRegistries.CONFIGURED_CARVER.getResourceKey(carver).orElseThrow());
     }
 
+    /**
+     * @param stage decoration stage
+     * @return all features registered for the given <code>stage</code>
+     */
     Iterable<Holder<PlacedFeature>> getFeatures(GenerationStep.Decoration stage);
 
+    /**
+     * @param stage carving stage
+     * @return all carvers registered for the given <code>stage</code>
+     */
     Iterable<Holder<ConfiguredWorldCarver<?>>> getCarvers(GenerationStep.Carving stage);
 }