+++ Common/src/main/java/fuzs/puzzleslib/api/networking/v3/MessageV3.java
@@ -3,7 +3,7 @@ package fuzs.puzzleslib.api.networking.v3;
 /**
  * Network message template providing a handler that runs when the message is received.
  */
-public interface Message<T> {
+public interface MessageV3<T> {
 
     /**
      * Create a handler for this message, usually {@link ClientMessageListener} or {@link ServerMessageListener}.