+++ Forge/src/main/java/fuzs/puzzleslib/network/ForgeNetworkHandler.java
@@ -22,10 +22,7 @@ import java.util.function.Supplier;
 
 /**
  * handler for network communications of all puzzles lib mods
- *
- * @deprecated migrate to {@link fuzs.puzzleslib.impl.networking.ForgeNetworkHandler}
  */
-@Deprecated(forRemoval = true)
 public class ForgeNetworkHandler implements NetworkHandler {
     /**
      * store network handlers created for a mod to avoid duplicate channels